notification added to all component

+++ client/src/Components/HostelVerificationDashboard.jsx
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import { DataGrid } from '@mui/x-data-grid';
 import { useFetch } from './../hooks/useFetch';
 import axios from 'axios';
-import { PromptContext } from '../context/PromptContext';
+
 import { NotificationContext } from './../context/NotificationContext';
 
 import DocumentImageModal from './DocumentImageModal';
@@ -25,34 +25,29 @@ const HostelVerificationDashboard = () => {
         setShowModal(true)
     }
 
-    const { status, dispatch } = useContext(PromptContext);
     const { dispatch: notificationDispatch } = useContext(NotificationContext)
     const verifyHostel = async (hostelId) => {
         const data = {
             "verified": true
         }
-        notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: "Hostel successfully Verified", status: "success" } })
-        // dispatch({ type: 'PROMPT_START', payload: { display: true, message: "Are you sure you want to verify this hostel?", purpose: "warning" } })
         try {
+          if (!window.confirm("Are you sure you want to verify this hostel?")) return
             const response = await axios.put(`http://localhost:5000/api/hostels/update/${hostelId}`, data, { withCredentials: true })
-            dispatch({ type: 'PROMPT_END' })
+            notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: "Hostel successfully Verified", status: "success" } })
             reFetchData()
         } catch (error) {
-            dispatch({ type: 'PROMPT_END' })
-            // notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: `${error.response.data.error}`, status: "error" } })
+            notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: "Hostel couldn't be verified", status: "error" } })
         }
     }
 
     const deleteHostel = async (hostelId) => {
-      dispatch({ type: 'PROMPT_START', payload: { display: true, message: "Are you sure you want to delete this hostel?", purpose: "alert" } })
         try {
+            if (!window.confirm("Are you sure you want to delete this hostel?")) return
             const response = await axios.delete(`http://localhost:5000/api/hostels/delete/${hostelId}`, { withCredentials: true })
-            dispatch({ type: 'PROMPT_END' })
             notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: "Hostel successfully deleted", status: "success" } })
             reFetchData()
         } catch (error) {
-            dispatch({ type: 'PROMPT_END' })
-            notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: `${error.response.data.error}`, status: "error" } })
+            notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: "Hostel couldn't be deleted", status: "error" } })
         }
     }
