+++ client/src/Components/UserVerificationDashboard.jsx
@@ -6,8 +6,6 @@ import { useFetch } from './../hooks/useFetch';
 import axios from 'axios';
 
 import DocumentImageModal from './DocumentImageModal';
-import PromptBar from './PromptBar';
-import { PromptContext } from '../context/PromptContext';
 import { NotificationContext } from './../context/NotificationContext';
 
 export const Wrapper = styled.div`
@@ -57,34 +55,31 @@ const UserVerificationDashboard = () => {
       setRows(data)
     }, [data])
 
-    const { status, dispatch } = useContext(PromptContext);
+
     const { dispatch: notificationDispatch } = useContext(NotificationContext)
     const verifyUser = async (userId) => {
       const data = {
         "usertype.is_verified": true
       }
-      dispatch({ type: 'PROMPT_START', payload: { display: true, message: "Are you sure you want to verify this user?", purpose: "warning" } })
       try {
+        if (!window.confirm("Are you sure you want to verify this user?")) return
         const response = await axios.put(`http://localhost:5000/api/users/update/${userId}`, data, { withCredentials: true })
-        dispatch({ type: 'PROMPT_END' })
         notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: "User successfully Verified", status: "success" } })
         reFetchData()
       } catch (error) {
-        dispatch({ type: 'PROMPT_END' })
-        notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: `${error.response.data.error}`, status: "error" } })
+        notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: "User couldn't be verified", status: "error" } })
       }
     }
 
     const deleteUser = async (userId) => {
-      dispatch({ type: 'PROMPT_START', payload: { display: true, message: "Are you sure you want to delete this user?", purpose: "alert" } })
+
       try {
+        if (!window.confirm("Are you sure you want to delete this user?")) return
         const response = await axios.delete(`http://localhost:5000/api/users/delete/${userId}`, { withCredentials: true })
-        dispatch({ type: 'PROMPT_END' })
         notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: "User successfully Deleted", status: "success" } })
         reFetchData()
       } catch (error) {
-        dispatch({ type: 'PROMPT_END' })
-        notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: `${error.response.data.error}`, status: "error" } })
+        notificationDispatch({ type: "NOTIFICATION_START", payload: { display: true, message: "User couldn't be deleted", status: "error" } })
       }
     }
